refactor : renderMoviesList 도메인 UI 분리
